feat: 토큰 생성 로직 분리
