+++ src/main/java/roomescape/auth/AuthService.java
@@ -1,35 +1,20 @@
 package roomescape.auth;
 
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
-import jakarta.annotation.PostConstruct;
-import java.security.Key;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import roomescape.auth.jwt.MemberTokenDto;
+import roomescape.auth.jwt.TokenService;
 import roomescape.member.Member;
 import roomescape.member.MemberDao;
 
 @Service
 public class AuthService {
 
-    private static final String NAME_CLAIM = "name";
-    private static final String EMAIL_CLAIM = "email";
-    private static final String ROLE_CLAIM = "role";
-
-    private final String secretKey;
     private final MemberDao memberDao;
+    private final TokenService tokenService;
 
-    private Key key;
-
-    public AuthService(@Value("${roomescape.auth.jwt.secret}") String secretKey,
-                       MemberDao memberDao) {
-        this.secretKey = secretKey;
+    public AuthService(MemberDao memberDao, TokenService tokenService) {
         this.memberDao = memberDao;
-    }
-
-    @PostConstruct
-    public void init() {
-        key = Keys.hmacShaKeyFor(secretKey.getBytes());
+        this.tokenService = tokenService;
     }
 
     public String loginWithEmailAndPassword(String email, String password) {
@@ -39,13 +24,7 @@ public class AuthService {
             throw new IllegalArgumentException("잘못된 이메일 또는 비밀번호입니다.");
         }
 
-        return Jwts.builder()
-                .setSubject(member.getId().toString())
-                .claim(NAME_CLAIM, member.getName())
-                .claim(EMAIL_CLAIM, member.getEmail())
-                .claim(ROLE_CLAIM, member.getRole())
-                .signWith(key)
-                .compact();
-
+        return tokenService.createToken(
+                new MemberTokenDto(member.getId(), member.getName(), member.getEmail(), member.getRole()));
     }
 }
