+++ src/main/java/roomescape/auth/jwt/MemberTokenDto.java
@@ -1,0 +1,9 @@
+package roomescape.auth.jwt;
+
+public record MemberTokenDto(
+        Long id,
+        String name,
+        String email,
+        String role
+) {
+}
